deepen: Mutation inlines edit-pipeline (shallow orchestrator deleted)
Move execPipeline + PipelineResult + snapshotIdFor (199 lines,
interface approx implementation) into mutation.ts. The single-edit
pipeline ordering (read, normalize, loadServed, applyOne, stableRehash,
drift) now lives in the deep seam that already owns the batch sequencer
and persist transaction. edit-pipeline.ts becomes a 6-line shim
re-exporting from mutation for backward compat.

No caller change needed: tool-edit already calls via mutation.applySingle.
Typecheck + 674 tests pass.
